+++ tests/modules/AccessMyVNA.py
@@ -20,31 +20,23 @@ import ctypes
 import logging
 logger = logging.getLogger(__name__)
 
+
+# load vna_wait_time_extra
 try:
-    from UISettings import config_default
+    import UISettings
+    config_default = UISettings.get_config()
+    settings_default = UISettings.get_settings()
 except:
     config_default = {}
+    settings_default = {}
 
-extra_time = None # initialize 
-if config_default:
-    usersettings_file = os.path.join(os.getcwd(), config_default['default_settings_file_name'])
-    if os.path.exists(usersettings_file):
-        try:
-            with open(usersettings_file, 'r') as f:
-                settings_user = json.load(f) # read user default settings
-                if 'vna_wait_time_extra' in settings_user:
-                    extra_time = settings_user['vna_wait_time_extra']
-            logger.info('use user settings')
-        except:
-            logger.warning('Error occured while loading {}\nuse default settings'.format(config_default['default_settings_file_name']))
-    else:
-        pass
-    del usersettings_file
-    if 'f' in locals():
-        del f
+# initialize extra_time by default
+extra_time = config_default.get('vna_wait_time_extra', 0.05) # in s. This extra time will be added to the calculated value
 
-if extra_time is None:
-    extra_time = config_default.get('vna_wait_time_extra', 0.05) # in s. This extra time will be added to the calculated value
+if 'vna_wait_time_extra' in settings_default:
+    extra_time = settings_default['vna_wait_time_extra']
+    logger.info('use user settings vna_wait_time_extra')
+# end load vna_wait_time_extra
 
 # try: # run from main
 #     from modules.retrying import retry
@@ -74,7 +66,8 @@ win_names = [
 # dll path
 # dll_path = r'./VNA/AccessMyVNA_v0.7/release/AccessMyVNAdll.dll'
 # dll_path = r'./VNA/AccessMyVNAv0.7_J/release/AccessMyVNAdll.dll'
-dll_path = r'./dll/AccessMyVNAdll.dll'
+
+dll_path = os.path.join(os.path.dirname(os.path.dirname(os.path.realpath(__file__))), 'dll', 'AccessMyVNAdll.dll')
 
 vna = WinDLL(dll_path, use_last_error=False) # this only works with AccessMyVNA
 # vna = OleDLL(r'AccessMyVNAdll.dll', use_last_error=True) # this only works with AccessMyVNA
@@ -1094,11 +1087,15 @@ class AccessMyVNA():
 
 
     # @retry(wait_fixed=wait_fixed, stop_max_attempt_number=stop_max_attempt_number, stop_max_delay=stop_max_delay, logger=True)
-    def SetDoubleArray(self, nWhat=0, nIndex=0, nArraySize=9, nData=[]):
+    def SetDoubleArray(self, nWhat=0, nIndex=0, nArraySize=9, nData=None):
         '''
         Set frequency nWhat = GET_SCAN_FREQ_DATA 0
         '''
         logger.info('MyVNASetDoubleArray') 
+        if nData is None: # nData has to be 
+            logger.info('no nData')
+            return 0, None
+
         if len(nData) == nArraySize: # check nData size
             if not isinstance(nData, np.ndarray): # check nData type
                 nData = np.array(nData)
@@ -1145,11 +1142,15 @@ class AccessMyVNA():
 
 
     # @retry(wait_fixed=wait_fixed, stop_max_attempt_number=stop_max_attempt_number, stop_max_delay=stop_max_delay, logger=True)
-    def SetIntegerArray(self, nWhat=5, nIndex=0, nArraySize=4, nData=[]):
+    def SetIntegerArray(self, nWhat=5, nIndex=0, nArraySize=4, nData=None):
         '''
         Set frequency nWhat = GET_SCAN_FREQ_DATA 0
         '''
         logger.info('MyVNASetIntegerArray') 
+        if nData is None: # no nData given
+            logger.info('no nData')
+            return 0, None
+            
         if len(nData) == nArraySize: # check nData size
             if not isinstance(nData, np.ndarray): # check nData type
                 nData = np.array(nData)
@@ -1398,9 +1399,9 @@ class AccessMyVNA():
         ret, nData = self.GetDoubleArray(nWhat=5, nIndex=0, nArraySize=2)
         logger.info('getADCChannel %s', nData)
 
-        if int(nData[0]) == 1:
+        if int(nData[1]) == 1:
             reflectchn = 1
-        elif int(nData[0]) == 2:
+        elif int(nData[1]) == 2:
             reflectchn =  2
         else: 
             reflectchn = None
@@ -1568,7 +1569,7 @@ class AccessMyVNA():
 if __name__ == '__main__':
     import time
     import logging
-    logging.basicConfig(level=logging.ERROR)
+    logging.basicConfig(level=logging.INFO)
     with AccessMyVNA() as vna:
         if vna.Init() == 0:
             print('open')
@@ -1576,7 +1577,8 @@ if __name__ == '__main__':
             print('false')
         pass
     print(vna)
-
+    print(vna._chn)
+    exit(0)
     with vna:
         print('start')
         # vna.SingleScan()
@@ -1587,10 +1589,10 @@ if __name__ == '__main__':
             if t - t0 > 1:
                 break
             ret, f, G = vna.GetScanData(nStart=0, nEnd=int(vna._nsteps-1), nWhata=-1, nWhatb=15)
-            print(i, t - t0, f[-1], G[-1])
+            # print(i, t - t0, f[-1], G[-1])
             time.sleep(0.01)
         t1 = time.time()
-        print(t1 - t0)
+        # print(t1 - t0)
         
         # vna.setADCChannel(reflectchn=2)
         # ret, delays = vna.GetIntegerArray(nWhat=5, nIndex=0, nArraySize=4)